+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 All notable changes to `NovaPoshta` will be documented in this file
 
+### 2020-10-30
+- [Add] `deleteScanSheet` in ScanSheet ($ScanSheetRefs)
+- [Add] `insertDocuments` in ScanSheet ($DocumentRefs)
+
 ### 2020-10-22
 - [Fix] `getWarehouses` in Address ($cityRef = null)
-
 
 ### 2020-10-27
 - [Add][ND] `getCatalogCounterparty` in Counterparty